desktop/requirements.txt
PyQt6
pandas
python-barcode
matplotlib
pytest
pytest-qt
argon2-cffi
python-dotenv
